Move pruning and quant API from paddleslim.dygraph to paddleslim
