+++ paddleslim/analysis/flops.py
@@ -18,7 +18,34 @@ from ..core import GraphWrapper, dygraph2program
 __all__ = ["flops", "dygraph_flops"]
 
 
-def flops(program, only_conv=True, detail=False):
+def flops(model, inputs=None, dtypes=None, only_conv=True, detail=False):
+    """
+    Compute the FLOPs of nn.Layer of paddle.Program.
+    Args:
+      model(paddle.nn.Layer|paddle.static.Program): The target model.
+      inputs(list): It is only used when model is instance of 'paddle.nn.Layer'. The dummy inputs used for 'model.forward'. It can be:
+                      1. list<int>|tuple<int>: means 'model.forward' accepts
+                         only one variable as argument and the shape of
+                         variable is 'inputs'.
+                      2. list<list<list>>: means 'model.forward' accepts multiple
+                         variables as arguments and the shapes of variables is 'inputs'.
+                      3. others: 'inputs' will be used as argument list by calling
+                         'model.forward(*inputs)'.
+      dtypes(str|list<str>): It only used when 'inputs' is shape or shapes that means
+                      data type of each input. None means all the inputs is 'float32'.
+                      Default: None.
+      only_conv(bool): Just return number of mul-adds in convolution and FC layer if `only_conv` is true.
+                         default: True.
+      detail(bool): Whether to return detail of each convolution layer.
+    """
+    if isinstance(model, paddle.static.Program):
+        return _static_flops(model, only_conv=only_conv, detail=detail)
+    elif isinstance(model, paddle.nn.Layer):
+        return dygraph_flops(
+            model, inputs, dtypes=dtypes, only_conv=only_conv, detail=detail)
+
+
+def _static_flops(program, only_conv=True, detail=False):
     """Get FLOPs of target graph.
 
     Args:
