+++ paddleslim/dygraph/__init__.py
@@ -1,24 +1,5 @@
-from . import var_group
-from .var_group import *
-from . import l1norm_pruner
-from .l1norm_pruner import *
-from . import pruner
-from .pruner import *
-from . import filter_pruner
-from .filter_pruner import *
-from . import l2norm_pruner
-from .l2norm_pruner import *
-from . import fpgm_pruner
-from .fpgm_pruner import *
-
 __all__ = []
-
-__all__ += var_group.__all__
-__all__ += l1norm_pruner.__all__
-__all__ += l2norm_pruner.__all__
-__all__ += fpgm_pruner.__all__
-__all__ += pruner.__all__
-__all__ += filter_pruner.__all__
-
 from .quant import *
 __all__ += quant.__all__
+from .prune import *
+__all__ += prune.__all__
