+++ paddleslim/dygraph/prune/l2norm_pruner.py
@@ -1,22 +1,23 @@
 import logging
 import numpy as np
 import paddle
-from ..common import get_logger
+from paddleslim.common import get_logger
 from .var_group import *
 from .pruning_plan import *
 from .filter_pruner import FilterPruner
 
-__all__ = ['L1NormFilterPruner']
+__all__ = ['L2NormFilterPruner']
 
 _logger = get_logger(__name__, logging.INFO)
 
 
-class L1NormFilterPruner(FilterPruner):
+class L2NormFilterPruner(FilterPruner):
     def __init__(self, model, inputs, sen_file=None):
-        super(L1NormFilterPruner, self).__init__(
+        super(L2NormFilterPruner, self).__init__(
             model, inputs, sen_file=sen_file)
 
     def cal_mask(self, var_name, pruned_ratio, group):
+        # find information of pruning on output channels
         for _item in group[var_name]:
             if _item['pruned_dims'] == [0]:
                 value = _item['value']
@@ -24,8 +25,10 @@ class L1NormFilterPruner(FilterPruner):
         reduce_dims = [
             i for i in range(len(value.shape)) if i not in pruned_dims
         ]
-        l1norm = np.mean(np.abs(value), axis=tuple(reduce_dims))
-        sorted_idx = l1norm.argsort()
+
+        # scores = np.mean(np.abs(value), axis=tuple(reduce_dims))
+        scores = np.sqrt(np.sum(np.square(value), axis=tuple(reduce_dims)))
+        sorted_idx = scores.argsort()
         pruned_num = int(round(len(sorted_idx) * pruned_ratio))
         pruned_idx = sorted_idx[:pruned_num]
         mask_shape = [value.shape[i] for i in pruned_dims]
