+++ paddleslim/dygraph/prune/var_group.py
@@ -2,9 +2,9 @@ import numpy as np
 import logging
 import paddle
 from paddle.fluid.dygraph import TracedLayer
-from ..core import GraphWrapper, dygraph2program
-from ..prune import collect_convs
-from ..common import get_logger
+from paddleslim.core import GraphWrapper, dygraph2program
+from paddleslim.prune import collect_convs
+from paddleslim.common import get_logger
 
 __all__ = ["VarGroup"]
