+++ tests/dygraph/test_prune.py
@@ -16,7 +16,7 @@ sys.path.append("../../")
 import unittest
 import paddle
 import paddle.fluid as fluid
-from paddleslim.dygraph import L1NormFilterPruner
+from paddleslim import L1NormFilterPruner
 from paddle.vision.models import mobilenet_v1, resnet50
 from paddleslim.prune import Pruner
